Moved the `TestContext` struct and its associated methods to a dedicated `test_context_test.go` file.
